+++ car-api/cars/cars-router.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const db = require('../data/dbConfig');
+const db = require('../../data/dbConfig');
 
 // Returns All Cars in the database
 router.get('/', async (req, res) => {
